Add priority support
Some mods want to be overridden by everyone without having to force
every single mod to specify a dependency. This allows the install
metadata to include a "priority" key. The packages with the lowest
priority gets loaded first, of course also taking into account
dependency chains.

My plan is to allow a package to include a default "priority" which the
user would be able to adjust on install, or after with a command. That
way we can allow users to easily swap around mods, but also allows mods
like usleep to be loaded first without having to have everythnigg
optdepend on it.